feat(partials): Optional (re)name partials functionality
We need more flexibility when it comes to naming partials and don't want to be constrained by our folder structure.

For example, we have

/components
    /button
        button.hbs
    /template
        otherTemplate.hbs
Currently, this names the partial button/button and template/otherTemplate.

This renaming function allows more flexibility, e.g., naming the two templates button and otherTemplate, ignoring the folder structure.